pi/requirements.txt
ipywidgets==8.1.2
jupyterlab==4.1.6
matplotlib==3.8.4
paho-mqtt==2.0.0
